refactor : VendingMachine.java
- 자판기에서 쓰이는 메서드들 클래스 이동.
