+++ src/main/java/vendingmachine/Application.java
@@ -26,45 +26,4 @@ public class Application {
         startVendingMachine(products);
     }
 
-
-    public static void printCurrentMachineCoin(){
-        System.out.println("자판기가 보유한 동전");
-        Coin[] eachCoin = Coin.values();
-        for(Coin c: eachCoin){
-            System.out.println(c.getAmount()+"원 - " + c.getCount()+"개");
-        }
-    }
-
-    public static void setRandomCountToEachCoin(){
-        int tempMachineOwnMoney = MACHINE_OWN_MONEY;
-        int selectedCoin;
-
-        while(tempMachineOwnMoney != 0) {
-            selectedCoin = Randoms.pickNumberInList(Coin.getCoinList());
-            if(tempMachineOwnMoney / selectedCoin > 0){
-                Coin.valueOf("COIN_" + selectedCoin).addCount();
-                tempMachineOwnMoney -= selectedCoin;
-            }
-        }
-    }
-
-    public static int inputMachineOwnMoney(){
-        String machineOwnMoney = "";
-        while(machineOwnMoney.isEmpty()) {
-            try {
-                machineOwnMoney = inputMachineOwnMoneyAndValidation();
-            } catch (IllegalArgumentException e){
-                System.out.println("[ERROR] 금액은 숫자여야 합니다.");
-            }
-        }
-        return Integer.parseInt(machineOwnMoney);
-    }
-
-    public static String inputMachineOwnMoneyAndValidation(){
-        System.out.println("자판기가 보유하고 있는 금액을 입력해 주세요.");
-        String machineOwnMoney = Console.readLine();
-        String regex = "^[0-9]{2,}$";
-            if (!Pattern.matches(regex, machineOwnMoney)) throw new IllegalArgumentException();
-        return machineOwnMoney;
-    }
 }
