+++ src/main/java/vendingmachine/VendingMachine.java
@@ -1,6 +1,10 @@
 package vendingmachine;
 
+import camp.nextstep.edu.missionutils.Console;
+import camp.nextstep.edu.missionutils.Randoms;
+
 import java.util.*;
+import java.util.regex.Pattern;
 
 import static vendingmachine.Application.USER_MONEY;
 import static vendingmachine.Validation.isUserMoneyAndMachineInventoryEnough;
@@ -56,5 +60,47 @@ public class VendingMachine {
             changes.put(coin.getAmount(), coinCount);
             USER_MONEY -= coinCount * coin.getAmount();
         }
+    }
+
+
+    public static void printCurrentMachineCoin(){
+        System.out.println("자판기가 보유한 동전");
+        Coin[] eachCoin = Coin.values();
+        for(Coin c: eachCoin){
+            System.out.println(c.getAmount()+"원 - " + c.getCount()+"개");
+        }
+    }
+
+    public static void setRandomCountToEachCoin(){
+        int tempMachineOwnMoney = MACHINE_OWN_MONEY;
+        int selectedCoin;
+
+        while(tempMachineOwnMoney != 0) {
+            selectedCoin = Randoms.pickNumberInList(Coin.getCoinList());
+            if(tempMachineOwnMoney / selectedCoin > 0){
+                Coin.valueOf("COIN_" + selectedCoin).addCount();
+                tempMachineOwnMoney -= selectedCoin;
+            }
+        }
+    }
+
+    public static int inputMachineOwnMoney(){
+        String machineOwnMoney = "";
+        while(machineOwnMoney.isEmpty()) {
+            try {
+                machineOwnMoney = inputMachineOwnMoneyAndValidation();
+            } catch (IllegalArgumentException e){
+                System.out.println("[ERROR] 금액은 숫자여야 합니다.");
+            }
+        }
+        return Integer.parseInt(machineOwnMoney);
+    }
+
+    public static String inputMachineOwnMoneyAndValidation(){
+        System.out.println("자판기가 보유하고 있는 금액을 입력해 주세요.");
+        String machineOwnMoney = Console.readLine();
+        String regex = "^[0-9]{2,}$";
+        if (!Pattern.matches(regex, machineOwnMoney)) throw new IllegalArgumentException();
+        return machineOwnMoney;
     }
 }
